Moving part of code to services according to specfic responsability

+++ src/main/java/com/sap/controllers/UserController.java
@@ -1,17 +1,17 @@
 package com.sap.controllers;
 
+import com.sap.dto.PassDTO;
+import com.sap.dto.UserDTO;
 import com.sap.models.*;
 import com.sap.service.*;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.view.RedirectView;
 
 import javax.annotation.Resource;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -31,60 +31,21 @@ public class UserController extends BaseController{
     /**
      * Method responsible for registering a new OWNER in the system. New member type users don't pass through this method.
      * @param model
-     * @param request
      * @return / redirects to the login page with a success or error message
      */
     @RequestMapping(value = "/register", method = RequestMethod.POST)
-    private ModelAndView addOwner(Model model, WebRequest request){
-        model.addAttribute("stay", true);
-        String username = request.getParameter("new_username");
-        String password = request.getParameter("new_password");
-        String confirm_pass = request.getParameter("confirm_password");
-
-        String error = "";
-        if(username.isEmpty() || password.isEmpty()){ error = "Username or password cannot be empty!"; }
-        if(userAlreadyExists(username)){ error = "Not possible to register the user in the system! Username already exists."; }
-        if(!new String(password).equals(confirm_pass)){ error = "Passwords doesn't match!"; }
-        if(!error.isEmpty()){
-            model.addAttribute("error", error);
-            return new ModelAndView("/login");
-        }
+    private ModelAndView addOwner(Model model, UserDTO user){
         try{
-            List<Role> roles = giveRoles(true);
-            this.createUser(username, password, createOwnersTeam(username), roles);
-            model.addAttribute("stay", false);
+            user.setRoles(giveRoles(true));
+            user.setTeam(createOwnersTeam(user.getUsername()));
+            this.userService.addUser(user);
             model.addAttribute("msg", "You've been registered successfully. Try to Log in!");
         }
-        catch (Exception e){
+        catch (IllegalArgumentException e){
+            model.addAttribute("stay", true);
             model.addAttribute("error", e.getMessage());
         }
         return new ModelAndView("/login");
-    }
-
-    /**
-     * Responsible for saving users in the database.
-     * @param username
-     * @param password
-     * @param team
-     */
-    private User createUser(String username, String password, Team team, List<Role> roles){
-        User user = new User();
-        user.setEnabled(true);
-        user.setUsername(username);
-        user.setPassword(passwordEncoder().encode(password));
-        user.setRoles(roles);
-        user.setTeam(team);
-        userService.addUser(user);
-        return user;
-    }
-
-    /**
-     * Verify if username already exists
-     * @param username
-     * @return boolean
-     */
-    public Boolean userAlreadyExists(String username){
-        return userService.getUserByName(username).getUsername() != null;
     }
 
     /**
@@ -130,11 +91,8 @@ public class UserController extends BaseController{
      */
     @RequestMapping("/user/delete/{id}")
     public RedirectView removeMember(@PathVariable("id") int id, Model model){
-        try{
-            userService.removeUser(id);
-        }catch (RuntimeException e){
-            return new RedirectView("/403");
-        }
+        if(this.userService.notAuthorized(id)){return new RedirectView("/403");}
+        userService.removeUser(id);
         return new RedirectView("/");
     }
 
@@ -165,31 +123,23 @@ public class UserController extends BaseController{
     /**
      * Responsible for adding new member. Action called through the owners interface
      * @param model
-     * @param request
      * @return success or error message in the add/edit page
      */
     @RequestMapping(value = "/user/add", method = RequestMethod.POST)
-    public String addMember(Model model, WebRequest request){
-        String username = request.getParameter("username");
-
-        if(username.isEmpty()){
-            model.addAttribute("error", "Username can't be empty!");
-            return "add-edit-user";
-        }
-        if(userAlreadyExists(username)){
-            model.addAttribute("error", "Not possible to add the member! Username already exists.");
-            return "add-edit-user";
-        }
-
-        try {
-            List<Role> roles = giveRoles(false);
-            User user = this.createUser(username, "password", this.getPrincipalUser().getTeam(), roles);
-            this.boundNewUserToPeriods(user);
+    public String addMember(Model model, UserDTO user){
+        try{
+            user.setRoles(giveRoles(false));
+            user.setTeam(this.getPrincipalUser().getTeam());
+            user.setPassword("password");
+            user.setConfirmPassword("password");
+            User new_user = this.userService.addUser(user);
+            this.boundNewUserToPeriods(new_user);
             model.addAttribute("msg", "New member registered successfully!");
         }
-        catch (Exception e){
+        catch (IllegalArgumentException e){
             model.addAttribute("error", e.getMessage());
         }
+
         return "add-edit-user";
     }
 
@@ -211,93 +161,39 @@ public class UserController extends BaseController{
         }
     }
 
-
     /**
      * Responsible for edit member type users
      * @param model
-     * @param request
      * @return success or error message in the add/edit page
      */
     @RequestMapping(value = "/user/edit", method = RequestMethod.POST)
-    public String saveEditedMember(Model model, WebRequest request) {
-
-        int id = Integer.parseInt(request.getParameter("id"));
-        int team_id = Integer.parseInt(request.getParameter("team"));
-        String username = request.getParameter("username");
-
-        User user = this.userService.getUserById(id);
-        Team team = this.teamService.getTeamById(team_id);
-
-        model.addAttribute("user", user);
-        model.addAttribute("teams", teamService.listTeams());
-
-        if (username.isEmpty()) {
-            model.addAttribute("error", "Username can't be empty!");
-            return "add-edit-user";
-        }
-        if (!new String(user.getUsername()).equals(username)){
-            if (userAlreadyExists(username)) {
-                model.addAttribute("error", "Not possible to edit member! Username already exists.");
-                return "add-edit-user";
-            }
-        }
+    public String saveEditedMember(Model model, UserDTO user) {
+        if(this.userService.notAuthorized(user.getId())){return "errors/403";}
         try
         {
-            user.setUsername(username);
-            user.setTeam(team);
+            user.setTeam(teamService.getTeamById(user.getTeamId()));
             userService.updateUser(user);
+            model.addAttribute("user", user);
+            model.addAttribute("teams", teamService.listTeams());
             model.addAttribute("msg", "Member edited successfully!");
         }
-        catch (Exception e){ model.addAttribute("error", e.getMessage()); }
+        catch (IllegalArgumentException e){ model.addAttribute("error", e.getMessage()); }
         return "add-edit-user";
-    }
-
-    /**
-     * Only redirects to the changepass user's page
-     * @return changepass page
-     */
-    @RequestMapping(value = "/changepass", method = RequestMethod.GET)
-    public String changePass(){
-        return "changepass";
     }
 
     /**
      * Allows user to change password when needed
      * @param model
-     * @param request
      * @return success or error message in the changepass page
      */
     @RequestMapping(value = "/changepass", method = RequestMethod.POST)
-    public String changePassdb(Model model, WebRequest request){
-
-        String oldpass = request.getParameter("old_password");
-        String newpass = request.getParameter("new_password");
-        String confirmpass = request.getParameter("confirm_password");
-
-        if(oldpass.isEmpty() || newpass.isEmpty() || confirmpass.isEmpty()){
-            model.addAttribute("error", "Fields can't be empty!");
-            return "changepass";
-        }
-
-        User principal = this.getPrincipalUser();
-        String userpass = principal.getPassword();
-
-        if(!passwordEncoder().matches(oldpass, userpass)){
-            model.addAttribute("error", "Current pass doesn't match!");
-            return "changepass";
-        }
-
-        if(!new String(newpass).equals(confirmpass)){
-            model.addAttribute("error", "Passwords doesn't match!");
-            return "changepass";
-        }
-
+    public String changePassdb(Model model, PassDTO pass){
         try{
-            principal.setPassword(passwordEncoder().encode(newpass));
-            userService.updateUser(principal);
+            pass.setUser(this.getPrincipalUser());
+            userService.updateUserPass(pass);
             model.addAttribute("msg", "Password changed successfully!");
         }
-        catch (Exception e){
+        catch (IllegalArgumentException e){
             model.addAttribute("error", e.getMessage());
         }
         return "changepass";
